+++ examples/basic/styleguide.config.js
@@ -3,6 +3,7 @@ var glob = require('glob');
 
 module.exports = {
 	title: 'React Style Guide Example',
+	defaultExample: true,
 	components: function() {
 		return glob.sync(path.resolve(__dirname, 'lib/components/**/*.js')).filter(function(module) {
 			return /\/[A-Z]\w*\.js$/.test(module);