+++ src/rsg-components/Components/Components.js
@@ -1,4 +1,4 @@
-import { Component, PropTypes } from 'react';
+import React, { Component, PropTypes } from 'react';
 import ReactComponent from 'rsg-components/ReactComponent';
 import Renderer from 'rsg-components/ReactComponent/Renderer';
 import Sections from 'rsg-components/Sections';
@@ -7,7 +7,7 @@ export default class Components extends Component {
 	static propTypes = {
 		highlightTheme: PropTypes.string.isRequired,
 		components: PropTypes.array.isRequired,
-		sections: PropTypes.array.isRequired
+		sections: PropTypes.array.isRequired,
 	};
 
 	renderComponents() {
@@ -15,7 +15,13 @@ export default class Components extends Component {
 		const ComponentRenderer = ReactComponent(Renderer);
 
 		return components.map((component) => {
-			return (<ComponentRenderer key={component.name} highlightTheme={highlightTheme} component={component} />);
+			return (
+				<ComponentRenderer
+					key={component.name}
+					highlightTheme={highlightTheme}
+					component={component}
+				/>
+			);
 		});
 	}