+++ src/rsg-components/Editor/Editor.js
@@ -2,42 +2,46 @@
 import '../../utils/codemirror-jsx';
 import 'codemirror/lib/codemirror.css';
 
-import _ from 'lodash';
-import { Component, PropTypes } from 'react';
+import React, { Component, PropTypes } from 'react';
 import debounce from 'lodash/debounce';
+import merge from 'lodash/merge';
 import Codemirror from 'react-codemirror';
 
 import s from './Editor.css';
 
-var cssRequire = require.context('codemirror/theme/', false, /^\.\/.*\.css$/);
+const codemirrorOptions = {
+	mode: 'jsx',
+	lineNumbers: false,
+	lineWrapping: true,
+	smartIndent: false,
+	matchBrackets: true,
+	viewportMargin: Infinity,
+};
 
-let UPDATE_DELAY = 100;
+const cssRequire = require.context('codemirror/theme/', false, /^\.\/.*\.css$/);
+
+let UPDATE_DELAY = 10;
 
 export default class Editor extends Component {
 	static propTypes = {
 		code: PropTypes.string.isRequired,
 		highlightTheme: PropTypes.string.isRequired,
-		onChange: PropTypes.func
-	};
-
-	static codemirrorOptions = {
-		mode: 'jsx',
-		lineNumbers: false,
-		lineWrapping: true,
-		smartIndent: false,
-		matchBrackets: true,
-		viewportMargin: Infinity
+		onChange: PropTypes.func,
 	};
 
 	constructor() {
 		super();
-		this._handleChange = debounce(this.handleChange.bind(this), UPDATE_DELAY);
+		this.handleChange = debounce(this.handleChange.bind(this), UPDATE_DELAY);
 	}
 
 	componentWillMount() {
 		let { highlightTheme } = this.props;
 
 		cssRequire(`./${highlightTheme}.css`);
+	}
+
+	shouldComponentUpdate() {
+		return false;
 	}
 
 	handleChange(newCode) {
@@ -49,11 +53,11 @@ export default class Editor extends Component {
 
 	render() {
 		let { highlightTheme } = this.props;
-		let options = _.merge({}, Editor.codemirrorOptions, {theme: highlightTheme});
+		let options = merge({}, codemirrorOptions, { theme: highlightTheme });
 
 		return (
 			<div className={s.root}>
-				<Codemirror value={this.props.code} onChange={this._handleChange} options={options}/>
+				<Codemirror value={this.props.code} onChange={this.handleChange} options={options} />
 			</div>
 		);
 	}