+++ src/rsg-components/Preview/Preview.js
@@ -10,13 +10,13 @@ import s from './Preview.css';
 export default class Preview extends Component {
 	static propTypes = {
 		code: PropTypes.string.isRequired,
-		evalInContext: PropTypes.func.isRequired
+		evalInContext: PropTypes.func.isRequired,
 	};
 
 	constructor() {
 		super();
 		this.state = {
-			error: null
+			error: null,
 		};
 		this.componentState = {};
 	}
@@ -33,7 +33,7 @@ export default class Preview extends Component {
 
 	compileCode(code) {
 		return transform(code, {
-			presets: ['es2015', 'react', 'stage-0']
+			presets: ['es2015', 'react', 'stage-0'],
 		}).code;
 	}
 
@@ -43,7 +43,7 @@ export default class Preview extends Component {
 		ReactDOM.unmountComponentAtNode(mountNode);
 
 		this.setState({
-			error: null
+			error: null,
 		});
 
 		let { code } = this.props;
@@ -111,7 +111,7 @@ export default class Preview extends Component {
 		catch (err) {
 			ReactDOM.unmountComponentAtNode(mountNode);
 			this.setState({
-				error: err.toString()
+				error: err.toString(),
 			});
 		}
 	}
@@ -123,9 +123,8 @@ export default class Preview extends Component {
 				<pre className={s.playgroundError}>{error}</pre>
 			);
 		}
-		else {
-			return null;
-		}
+
+		return null;
 	}
 
 	render() {