+++ src/server.js
@@ -1,14 +1,19 @@
+/* eslint-disable no-var */
+
 var express = require('express');
 var webpack = require('webpack');
-var makeWebpackConfig = require('./make-webpack-config');
+
 var config = require('./utils/config');
+config.initialize(); // we need to initialize config before requiring anything else from the project
+
+var makeWebpackConfig = require('./make-webpack-config');
 
 module.exports = function server(callback) {
 	var app = express();
 	var compiler = webpack(makeWebpackConfig('development'));
 
 	app.use(require('webpack-dev-middleware')(compiler, {
-		noInfo: true
+		noInfo: true,
 	}));
 
 	app.use(require('webpack-hot-middleware')(compiler));